Refactor PhotoGrid into generic Grid component
Extract Photo related component, code and rendering to PhotoGridScreen
